requirements.txt
numpy
pandas
Pillow
tqdm
torch
scikit-learn